Closes: https://github.com/goharbor/harbor/issues/16456

Signed-off-by: chlins <chenyuzh@vmware.com>

Comparison: https://github.com/goharbor/harbor/pull/16546#issuecomment-1070516463