@chlins 2022-03-17
Performance comparison after using harbor session provider.

Test tool: https://github.com/rakyll/hey

### Case 1

*No extra business logic, simple API call.*

Script: `hey -c 2000 -z 1m http://harbor.domain/api/version`

before
```bash
Summary:
  Total:        61.3372 secs
  Slowest:      7.6496 secs
  Fastest:      0.0042 secs
  Average:      0.3689 secs
  Requests/sec: 5393.5916

  Total data:   32520442 bytes
  Size/request: 98 bytes

Response time histogram:
  0.004 [1]     |
  0.769 [295083]        |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  1.533 [12464] |■■
  2.298 [9724]  |■
  3.062 [4105]  |■
  3.827 [5172]  |■
  4.591 [2422]  |
  5.356 [1293]  |
  6.121 [471]   |
  6.885 [76]    |
  7.650 [17]    |


Latency distribution:
  10% in 0.0746 secs
  25% in 0.1147 secs
  50% in 0.1319 secs
  75% in 0.1639 secs
  90% in 0.9772 secs
  95% in 1.9432 secs
  99% in 4.1307 secs

Details (average, fastest, slowest):
  DNS+dialup:   0.0003 secs, 0.0042 secs, 7.6496 secs
  DNS-lookup:   0.0000 secs, 0.0000 secs, 0.0000 secs
  req write:    0.0000 secs, 0.0000 secs, 0.0317 secs
  resp wait:    0.3684 secs, 0.0042 secs, 7.6496 secs
  resp read:    0.0000 secs, 0.0000 secs, 0.0263 secs

Status code distribution:
  [200] 317263 responses
  [503] 13565 responses
```

after
```bash
Summary:
  Total:        60.1015 secs
  Slowest:      1.0474 secs
  Fastest:      0.0027 secs
  Average:      0.1662 secs
  Requests/sec: 12009.1263

  Total data:   13713554 bytes
  Size/request: 19 bytes

Response time histogram:
  0.003 [1]     |
  0.107 [181988]        |■■■■■■■■■■■■■■■■■■■■
  0.212 [369184]        |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  0.316 [123155]        |■■■■■■■■■■■■■
  0.421 [34554] |■■■■
  0.525 [9721]  |■
  0.630 [2123]  |
  0.734 [737]   |
  0.838 [301]   |
  0.943 [0]     |
  1.047 [2]     |


Latency distribution:
  10% in 0.0925 secs
  25% in 0.1070 secs
  50% in 0.1337 secs
  75% in 0.2062 secs
  90% in 0.2834 secs
  95% in 0.3378 secs
  99% in 0.4620 secs

Details (average, fastest, slowest):
  DNS+dialup:   0.0001 secs, 0.0027 secs, 1.0474 secs
  DNS-lookup:   0.0000 secs, 0.0000 secs, 0.0000 secs
  req write:    0.0000 secs, 0.0000 secs, 0.0392 secs
  resp wait:    0.1660 secs, 0.0026 secs, 1.0470 secs
  resp read:    0.0001 secs, 0.0000 secs, 0.0403 secs

Status code distribution:
  [200] 721766 responses
```

In conclusion,  QPS improvement is about **2.2X** than before and P99 latency is **9X** faster than before.

### Case 2

*Most commonly used harbor API call.*

Script: `hey -c 2000 -z 1m -H "Authorization: Basic YWRtaW46SGFyYm9yMTIzNDUK" http://harbor.domain/api/v2.0/projects/library/repositories/busybox/artifacts/latest`

before
```bash
Summary:
  Total:        62.0304 secs
  Slowest:      6.7418 secs
  Fastest:      0.0201 secs
  Average:      0.8670 secs
  Requests/sec: 2250.8785

  Total data:   163334703 bytes
  Size/request: 1169 bytes

Response time histogram:
  0.020 [1]     |
  0.692 [100077]        |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  1.364 [15208] |■■■■■■
  2.037 [7712]  |■■■
  2.709 [8787]  |■■■■
  3.381 [4461]  |■■
  4.053 [2153]  |■
  4.725 [863]   |
  5.397 [282]   |
  6.070 [62]    |
  6.742 [17]    |


Latency distribution:
  10% in 0.3512 secs
  25% in 0.4233 secs
  50% in 0.5248 secs
  75% in 0.7498 secs
  90% in 2.2207 secs
  95% in 2.8043 secs
  99% in 3.9978 secs

Details (average, fastest, slowest):
  DNS+dialup:   0.0006 secs, 0.0201 secs, 6.7418 secs
  DNS-lookup:   0.0000 secs, 0.0000 secs, 0.0000 secs
  req write:    0.0001 secs, 0.0000 secs, 0.0224 secs
  resp wait:    0.8662 secs, 0.0201 secs, 6.7417 secs
  resp read:    0.0000 secs, 0.0000 secs, 0.0186 secs

Status code distribution:
  [200] 133028 responses
  [503] 6595 responses
```

after
```bash
Summary:
  Total:        61.0273 secs
  Slowest:      4.0263 secs
  Fastest:      0.1018 secs
  Average:      0.5628 secs
  Requests/sec: 3506.3822

  Total data:   242017035 bytes
  Size/request: 1131 bytes

Response time histogram:
  0.102 [1]     |
  0.494 [100964]        |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  0.887 [97765] |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  1.279 [10744] |■■■■
  1.672 [1599]  |■
  2.064 [1557]  |■
  2.456 [1116]  |
  2.849 [200]   |
  3.241 [21]    |
  3.634 [12]    |
  4.026 [6]     |

Latency distribution:
  10% in 0.3383 secs
  25% in 0.4083 secs
  50% in 0.5055 secs
  75% in 0.6399 secs
  90% in 0.8134 secs
  95% in 0.9707 secs
  99% in 1.9378 secs

Details (average, fastest, slowest):
  DNS+dialup:   0.0006 secs, 0.1018 secs, 4.0263 secs
  DNS-lookup:   0.0000 secs, 0.0000 secs, 0.0000 secs
  req write:    0.0001 secs, 0.0000 secs, 0.0596 secs
  resp wait:    0.5620 secs, 0.0688 secs, 4.0262 secs
  resp read:    0.0000 secs, 0.0000 secs, 0.0197 secs

Status code distribution:
  [200] 213985 responses
```

In conclusion,  QPS improvement is about **1.5X** than before and P99 latency is **2X** faster than before.

### Case 3

*Distribution V2 manifest API call.*

Script: `hey -c 2000 -z 1m -H "Authorization: Basic YWRtaW46SGFyYm9yMTIzNDUK" http://harbor.domain/v2/library/busybox/manifests/latest`

before
```bash
Summary:
  Total:        62.4289 secs
  Slowest:      8.7075 secs
  Fastest:      0.0161 secs
  Average:      2.2138 secs
  Requests/sec: 885.4873

  Total data:   22974038 bytes
  Size/request: 415 bytes

Response time histogram:
  0.016 [1]     |
  0.885 [4396]  |■■■■■■■■
  1.754 [11593] |■■■■■■■■■■■■■■■■■■■■
  2.624 [22900] |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  3.493 [11691] |■■■■■■■■■■■■■■■■■■■■
  4.362 [3503]  |■■■■■■
  5.231 [957]   |■■
  6.100 [150]   |
  6.969 [56]    |
  7.838 [26]    |
  8.708 [7]     |


Latency distribution:
  10% in 1.0159 secs
  25% in 1.6265 secs
  50% in 2.1889 secs
  75% in 2.7485 secs
  90% in 3.4020 secs
  95% in 3.7632 secs
  99% in 4.9099 secs

Details (average, fastest, slowest):
  DNS+dialup:   0.0018 secs, 0.0161 secs, 8.7075 secs
  DNS-lookup:   0.0000 secs, 0.0000 secs, 0.0000 secs
  req write:    0.0001 secs, 0.0000 secs, 0.0152 secs
  resp wait:    2.2112 secs, 0.0160 secs, 8.7075 secs
  resp read:    0.0001 secs, 0.0000 secs, 0.0220 secs

Status code distribution:
  [200] 43594 responses
  [502] 11686 responses
```

after
```bash
Summary:
  Total:        62.6886 secs
  Slowest:      8.3946 secs
  Fastest:      0.0332 secs
  Average:      2.1092 secs
  Requests/sec: 924.7620

  Total data:   18182554 bytes
  Size/request: 313 bytes

Response time histogram:
  0.033 [1]     |
  0.869 [4264]  |■■■■■■■
  1.705 [14318] |■■■■■■■■■■■■■■■■■■■■■■■
  2.542 [24612] |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  3.378 [9046]  |■■■■■■■■■■■■■■■
  4.214 [4467]  |■■■■■■■
  5.050 [1055]  |■■
  5.886 [145]   |
  6.722 [50]    |
  7.558 [9]     |
  8.395 [5]     |


Latency distribution:
  10% in 1.0335 secs
  25% in 1.5526 secs
  50% in 2.0387 secs
  75% in 2.5553 secs
  90% in 3.3706 secs
  95% in 3.8346 secs
  99% in 4.4561 secs

Details (average, fastest, slowest):
  DNS+dialup:   0.0016 secs, 0.0332 secs, 8.3946 secs
  DNS-lookup:   0.0000 secs, 0.0000 secs, 0.0000 secs
  req write:    0.0001 secs, 0.0000 secs, 0.0237 secs
  resp wait:    2.1066 secs, 0.0331 secs, 8.3945 secs
  resp read:    0.0000 secs, 0.0000 secs, 0.0163 secs

Status code distribution:
  [200] 34502 responses
  [502] 23470 responses
```

In conclusion,  QPS and P95 latency have no obvious improvement due to the bottleneck in distribution.