@chlins 2022-03-24
> Should we move core/session to lib/session to follow the convention?

@stonezdj I don't think so, the pr session is only designed and used for core.